core data tests

+++ WeatherApp/Services/Storage/CoreDataStorage.swift
@@ -10,7 +10,7 @@ import UIKit
 import CoreData
 
 protocol CoreDataStorage {
-    func fetchList() -> [CWEntity]?
+    func fetchList() -> [CWEntity]
     func addToList(city: CWEntity)
     func deleteFromList(index: Int)
     func updateList(index: Int, entity: CWEntity)
@@ -20,17 +20,20 @@ final class CoreDataStorageImp: CoreDataStorage  {
     
     let context = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
     
-    func fetchList() -> [CWEntity]? {
-        guard let request = try? context.fetch(Cities.fetchRequest()) else {return nil}
+    func fetchList() -> [CWEntity] {
+        var array = [CWEntity]()
+        guard let request = try? context.fetch(Cities.fetchRequest()) else {
+            return array
+        }
+        
         let sortRequest = request.sorted{$0.date! < $1.date!}
         let decoder = JSONDecoder()
-        var array = [CWEntity]()
         sortRequest.forEach { row in
             guard let data: Data = row.city else {return}
             guard let city = try? decoder.decode(CWEntity.self, from: data) else {return}
             array.append(city)
         }
-        return array.count != 0 ? array : nil
+        return array//.count != 0 ? array : nil
     }
     
     func addToList(city: CWEntity){
@@ -49,22 +52,27 @@ final class CoreDataStorageImp: CoreDataStorage  {
     func deleteFromList(index: Int) {
         guard let request = try? context.fetch(Cities.fetchRequest()) else {return}
         let sortRequest = request.sorted{$0.date! < $1.date!}
-        let city = sortRequest[index]
-        context.delete(city)
-        context.perform {
-            try? self.context.save()
+        if index < sortRequest.count {
+            let city = sortRequest[index]
+            context.delete(city)
+            context.perform {
+                try? self.context.save()
+            }
         }
     }
     
     func updateList(index: Int, entity: CWEntity) {
         guard let request = try? context.fetch(Cities.fetchRequest()) else {return}
         let sortRequest = request.sorted{$0.date! < $1.date!}
-        let row = sortRequest[index]
-        let encoder = JSONEncoder()
-        guard let data: Data = try? encoder.encode(entity) else {return}
-        row.city = data
-        context.perform {
-            try? self.context.save()
+        
+        if index < sortRequest.count {
+            let row = sortRequest[index]
+            let encoder = JSONEncoder()
+            guard let data: Data = try? encoder.encode(entity) else {return}
+            row.city = data
+            context.perform {
+                try? self.context.save()
+            }
         }
     }
 }
